Find DNLA profiles by resource attributes
Extracted ConfigSetup classes to subdir
Add new ConfigSetupVector to store any type of key-value pairs
