+++ src/config/config_manager.h
@@ -90,6 +90,10 @@ public:
     /// \param option option to retrieve.
     std::map<std::string, std::string> getDictionaryOption(config_option_t option) const override;
 
+    /// \brief returns a config option of type vector
+    /// \param option option to retrieve.
+    std::vector<std::vector<std::pair<std::string, std::string>>> getVectorOption(config_option_t option) const override;
+
     /// \brief returns a config option of type array of string
     /// \param option option to retrieve.
     std::vector<std::string> getArrayOption(config_option_t option) const override;
