+++ src/config/config_setup.h
@@ -538,7 +538,7 @@ protected:
     /// </some-section>
     ///
     /// This function will create a dictionary with the following
-    /// key:value paris: "1":"2", "3":"4"
+    /// key:value pairs: "1":"2", "3":"4"
     bool createOptionFromNode(const pugi::xml_node& optValue, std::map<std::string, std::string>& result) const;
 
     bool updateItem(std::size_t i, const std::string& optItem, const std::shared_ptr<Config>& config, const std::shared_ptr<DictionaryOption>& value, const std::string& optKey, const std::string& optValue, const std::string& status = "") const;
@@ -582,6 +582,61 @@ public:
     std::map<std::string, std::string> getXmlContent(const pugi::xml_node& optValue);
 
     std::shared_ptr<ConfigOption> newOption(const std::map<std::string, std::string>& optValue);
+
+    std::string getCurrentValue() const override { return {}; }
+};
+
+class ConfigVectorSetup : public ConfigSetup {
+protected:
+    bool notEmpty = false;
+    bool itemNotEmpty = false;
+    bool tolower = false;
+    std::vector<std::vector<std::pair<std::string, std::string>>> defaultEntries;
+
+    /// \brief Creates a vector from an XML nodeset.
+    /// \param optValue starting element of the nodeset.
+    ///
+    /// The basic idea is the following:
+    /// You have a piece of XML that looks like this
+    /// <some-section>
+    ///    <map from="1" via="3" to="2"/>
+    ///    <map from="3" to="4"/>
+    /// </some-section>
+    ///
+    /// This function will create a vector with the following
+    /// list: { { "1", "3", "2" }, {"3", "", "4"}
+    bool createOptionFromNode(const pugi::xml_node& optValue, std::vector<std::vector<std::pair<std::string, std::string>>>& result) const;
+
+    bool updateItem(std::size_t i, const std::string& optItem, const std::shared_ptr<Config>& config, const std::shared_ptr<VectorOption>& value, const std::string& optValue, const std::string& status = "") const;
+
+public:
+    config_option_t nodeOption {};
+    std::vector<config_option_t> optionList {};
+
+    explicit ConfigVectorSetup(config_option_t option, const char* xpath, const char* help,
+        config_option_t nodeOption, std::vector<config_option_t> optionList,
+        bool notEmpty = false, bool itemNotEmpty = false, bool required = false, std::vector<std::vector<std::pair<std::string, std::string>>> defaultEntries = {})
+        : ConfigSetup(option, xpath, help, required && defaultEntries.empty())
+        , notEmpty(notEmpty)
+        , itemNotEmpty(itemNotEmpty)
+        , defaultEntries(std::move(defaultEntries))
+        , nodeOption(nodeOption)
+        , optionList(optionList)
+    {
+    }
+
+    std::string getTypeString() const override { return "List"; }
+
+    void makeOption(const pugi::xml_node& root, const std::shared_ptr<Config>& config, const std::map<std::string, std::string>* arguments = nullptr) override;
+
+    bool updateDetail(const std::string& optItem, std::string& optValue, const std::shared_ptr<Config>& config, const std::map<std::string, std::string>* arguments = nullptr) override;
+
+    std::string getItemPath(int index = 0, config_option_t propOption = CFG_MAX, config_option_t propOption2 = CFG_MAX, config_option_t propOption3 = CFG_MAX, config_option_t propOption4 = CFG_MAX) const override;
+    std::string getItemPath(int index, const std::string propOption) const;
+
+    std::vector<std::vector<std::pair<std::string, std::string>>> getXmlContent(const pugi::xml_node& optValue);
+
+    std::shared_ptr<ConfigOption> newOption(const std::vector<std::vector<std::pair<std::string, std::string>>>& optValue);
 
     std::string getCurrentValue() const override { return {}; }
 };
