+++ src/upnp_xml.h
@@ -112,7 +112,7 @@ protected:
     std::string entrySeparator;
     bool multiValue {};
     std::map<std::string, std::string> ctMappings;
-    std::map<std::string, std::string> profMappings;
+    std::vector<std::vector<std::pair<std::string, std::string>>> profMappings;
     std::map<std::string, std::string> transferMappings;
 
     /// \brief Holds a part of path and bool which says if we need to append the resource
@@ -132,6 +132,7 @@ protected:
     static std::string renderExtension(const std::string& contentType, const fs::path& location);
     static void addField(pugi::xml_node& entry, const std::string& key, const std::string& val);
     void addPropertyList(pugi::xml_node& result, const std::vector<std::pair<std::string, std::string>>& meta, const std::map<std::string, std::string>& auxData, config_option_t itemProps, config_option_t nsProp);
+    std::string findDlnaProfile(const std::shared_ptr<CdsResource>& res, const std::string& contentType);
     std::string dlnaProfileString(const std::shared_ptr<CdsResource>& res, const std::string& contentType, bool formatted = true);
 };
 #endif // __UPNP_XML_H__
