+++ test/database/mysql_config_fake.h
@@ -53,6 +53,7 @@ public:
     int getIntOption(config_option_t option) const override { return 0; }
     bool getBoolOption(config_option_t option) const override { return false; }
     std::map<std::string, std::string> getDictionaryOption(config_option_t option) const override { return {}; }
+    std::vector<std::vector<std::pair<std::string, std::string>>> getVectorOption(config_option_t option) const override { return {}; }
     std::vector<std::string> getArrayOption(config_option_t option) const override { return {}; }
     std::shared_ptr<AutoscanList> getAutoscanListOption(config_option_t option) const override { return nullptr; }
     std::shared_ptr<ClientConfigList> getClientConfigListOption(config_option_t option) const override { return nullptr; }
